Move TokenSwaps to project traits
